One modification and one addition was made:
MODIFICATION
============
We got rid of the three step process of question paper generation as earlier.
Now, it's done in two steps.
All mention of question papers go away.
Assessment instruments, which were the intermediate artefact, is now the final
artefact.

ENHANCEMENT
===========
We added a backup directory. This is facilitate testing the
application post assessment instrument generation + quiz execution step.
This was needed because the gen.py generates a random set of
assessment instruments which invalidates the earlier roll-number to assessment
instrument mapping.
backup/restore.sh restores the earlier setting.